+++ grafana_wtf/model.py
@@ -1,0 +1,31 @@
+import dataclasses
+from typing import List
+
+from munch import Munch
+from collections import OrderedDict
+from urllib.parse import urljoin
+
+
+@dataclasses.dataclass
+class DatasourceBreakdownItem:
+    datasource: Munch
+    used_in: List[Munch]
+    grafana_url: str
+
+    def format_compact(self):
+        dsshort = OrderedDict(
+            name=self.datasource.name,
+            type=self.datasource.type,
+            url=self.datasource.url,
+        )
+        item = OrderedDict(datasource=dsshort)
+        for dashboard in self.used_in:
+            item.setdefault("dashboards", [])
+            dbshort = OrderedDict(
+                title=dashboard.dashboard.title,
+                uid=dashboard.dashboard.uid,
+                path=dashboard.meta.url,
+                url=urljoin(self.grafana_url, dashboard.meta.url),
+            )
+            item["dashboards"].append(dbshort)
+        return item